Move tests to their own file too
